+++ PSFramework/changelog.md
@@ -9,6 +9,7 @@
  - New: Logging Provider for gelf / graylog
  - Upd: PsfValidateScript can now consume stored scriptblocks
  - Upd: PsfValidateScript will now understand both $_ and $args[0]
+ - Upd: PsfValidateSet now supports localized strings using the `ErrorString` property.
  - Fix: PsfValidateSet fails on completion scriptblock with whitespace value
 
 ## 0.10.28.144 : 2018-10-28